housekeeping: refactor FlashMessage alert && move variable and funtion into store && update type
