+++ apps/web/app/[id]/types.ts
@@ -2,3 +2,8 @@ export interface FormEntry {
   name: string;
   label: string;
 }
+
+export interface NotificationType {
+  message: string;
+  severity: 'success' | 'warning';
+}
